+++ src/pem/__init__.py
@@ -6,6 +6,7 @@ from ._core import (
     CertificateRequest,
     CertificateRevocationList,
     DHParameters,
+    DSAPrivateKey,
     ECPrivateKey,
     Key,
     OpenSSHPrivateKey,
@@ -26,7 +27,7 @@ except ImportError:
     twisted = None  # type: ignore
 
 
-__version__ = "20.2.0.dev0"
+__version__ = "21.1.0.dev0"
 __author__ = "Hynek Schlawack"
 __license__ = "MIT"
 __description__ = "Easy PEM file parsing in Python."
@@ -41,6 +42,7 @@ __all__ = [
     "CertificateRequest",
     "CertificateRevocationList",
     "DHParameters",
+    "DSAPrivateKey",
     "ECPrivateKey",
     "Key",
     "OpenSSHPrivateKey",