+++ src/pem/_core.py
@@ -170,6 +170,16 @@ class ECPrivateKey(PrivateKey):
     """
 
 
+class DSAPrivateKey(PrivateKey):
+    """
+    A private DSA key.
+
+    Also private DSA key in OpenSSH legacy PEM format.
+
+    .. versionadded:: 21.1.0
+    """
+
+
 class DHParameters(AbstractPEMObject):
     """
     Diffie-Hellman parameters for DHE.
@@ -209,6 +219,7 @@ _PEM_TO_CLASS = {
     b"PUBLIC KEY": PublicKey,
     b"ENCRYPTED PRIVATE KEY": PrivateKey,
     b"OPENSSH PRIVATE KEY": OpenSSHPrivateKey,
+    b"DSA PRIVATE KEY": DSAPrivateKey,
     b"RSA PRIVATE KEY": RSAPrivateKey,
     b"RSA PUBLIC KEY": RSAPublicKey,
     b"EC PRIVATE KEY": ECPrivateKey,